edit device and category

+++ src/categories/categories.controller.js
@@ -1,30 +1,36 @@
 const mongoose = require("mongoose");
 
 const Category = require("../../db/config/categories.model");
+const manageDevice = require("../../db/config/devicemanage.model");
 
 // create Category
 
 const createCategory = async (req, res) => {
   try {
     const { name } = req.body;
-
     console.log("category data", req.body);
-
     // const icon = req.file ? req?.file?.filename : null;
     const icon = req.files ? req.files.map((file) => file.filename) : null;
+    const cat = await Category.findOne({ name });
+    if (cat) {
+      res.status(400).json({
+        success: false,
+        error: "Category already listed by this name",
+      });
+    } else {
+      const newCategory = new Category({
+        name,
+        icon,
+      });
 
-    const newCategory = new Category({
-      name,
-      icon,
-    });
+      await newCategory.save();
 
-    await newCategory.save();
-
-    res.status(201).json({
-      success: true,
-      message: "Category created successfully",
-      data: newCategory,
-    });
+      res.status(201).json({
+        success: true,
+        message: "Category created successfully",
+        data: newCategory,
+      });
+    }
   } catch (error) {
     console.error(error);
     res.status(500).json({ success: false, error: "Internal Server Error" });
@@ -100,18 +106,27 @@ const getCategoryById = async (req, res) => {
 const updateCategory = async (req, res) => {
   try {
     const categoryId = req.params.categoryId;
-
+    const data = req.body;
     if (!mongoose.Types.ObjectId.isValid(categoryId)) {
       return res
         .status(400)
         .json({ success: false, error: "Invalid category ID" });
     }
 
-    const updatedCategory = await Category.findByIdAndUpdate(
-      categoryId,
-      { $set: req.body },
-      { new: true, runValidators: true }
-    );
+    const icon = req.files
+      .filter((item) => item.fieldname === "icon")
+      .map((file) => file.filename);
+    if (icon.length > 0) {
+      console.log(icon);
+      data.icon = icon;
+    } else {
+      data.icon = [data.icon];
+    }
+
+    const updatedCategory = await Category.findByIdAndUpdate(categoryId, data, {
+      new: true,
+      runValidators: true,
+    });
 
     if (!updatedCategory) {
       return res
@@ -130,22 +145,29 @@ const updateCategory = async (req, res) => {
 const deleteCategory = async (req, res) => {
   try {
     const categoryId = req.params.categoryId;
-
     if (!mongoose.Types.ObjectId.isValid(categoryId)) {
       return res
         .status(400)
         .json({ success: false, error: "Invalid category ID" });
     }
-
-    const deletedCategory = await Category.findByIdAndDelete(categoryId);
-
-    if (!deletedCategory) {
-      return res
-        .status(404)
-        .json({ success: false, error: "Category not found" });
+    const devices = await manageDevice.find({ category: categoryId });
+    console.log(devices);
+    if (devices.length == 0) {
+      const deletedCategory = await Category.findByIdAndDelete(categoryId);
+      res.status(200).json({ success: true, data: deletedCategory });
+    } else {
+      return res.status(400).json({
+        success: false,
+        error: "cannot delete category because device listed on this category",
+      });
     }
+    // if (!deletedCategory) {
+    //   return res
+    //     .status(404)
+    //     .json({ success: false, error: "Category not found" });
+    // }
 
-    res.status(200).json({ success: true, data: deletedCategory });
+    // res.status(200).json({ success: true, data: deletedCategory });
   } catch (error) {
     console.error(error);
     res.status(500).json({ success: false, error: "Internal Server Error" });
